Add files via upload
Here is an example how we can collect data from url using API. Here, the context is about survey on popular jobs trending and future.We use the get method of requests http module to access necessary data from the dataset. At the last part the output is saved to saved into an excel file.